+++ spec/active_force/active_query_spec.rb
@@ -63,6 +63,24 @@ describe ActiveForce::ActiveQuery do
       expect(active_query.to_s).to end_with("(Field__c = 'hello')")
     end
 
+    it "formats as YYYY-MM-DDThh:mm:ss-hh:mm and does not enclose in quotes if it's a DateTime" do
+      value = DateTime.now
+      active_query.where(field: value)
+      expect(active_query.to_s).to end_with("(Field__c = #{value.iso8601})")
+    end
+
+    it "formats as YYYY-MM-DDThh:mm:ss-hh:mm and does not enclose in quotes if it's a Time" do
+      value = Time.now
+      active_query.where(field: value)
+      expect(active_query.to_s).to end_with("(Field__c = #{value.iso8601})")
+    end
+
+    it "formats as YYYY-MM-DD and does not enclose in quotes if it's a Date" do
+      value = Date.today
+      active_query.where(field: value)
+      expect(active_query.to_s).to end_with("(Field__c = #{value.iso8601})")
+    end
+
     it "puts NULL when a field is set as nil" do
       active_query.where field: nil
       expect(active_query.to_s).to end_with("(Field__c = NULL)")
@@ -91,6 +109,24 @@ describe ActiveForce::ActiveQuery do
         expect(active_query.to_s).to eq("SELECT Id FROM table_name WHERE (Field__c = 123 AND Other_Field__c = 321 AND Name = 'Bob')")
       end
 
+      it 'formats as YYYY-MM-DDThh:mm:ss-hh:mm and does not enclose in quotes if value is a DateTime' do
+        value = DateTime.now
+        active_query.where('Field__c > ?', value)
+        expect(active_query.to_s).to end_with("(Field__c > #{value.iso8601})")
+      end
+
+      it 'formats as YYYY-MM-DDThh:mm:ss-hh:mm and does not enclose in quotes if value is a Time' do
+        value = Time.now
+        active_query.where('Field__c > ?', value)
+        expect(active_query.to_s).to end_with("(Field__c > #{value.iso8601})")
+      end
+
+      it 'formats as YYYY-MM-DD and does not enclose in quotes if value is a Date' do
+        value = Date.today
+        active_query.where('Field__c > ?', value)
+        expect(active_query.to_s).to end_with("(Field__c > #{value.iso8601})")
+      end
+
       it 'complains when there given an incorrect number of bind parameters' do
         expect{
           active_query.where('Field__c = ? AND Other_Field__c = ? AND Name = ?', 123, 321)
@@ -106,6 +142,24 @@ describe ActiveForce::ActiveQuery do
         it 'accepts nil bind parameters' do
           active_query.where('Field__c = :field', field: nil)
           expect(active_query.to_s).to eq("SELECT Id FROM table_name WHERE (Field__c = NULL)")
+        end
+
+        it 'formats as YYYY-MM-DDThh:mm:ss-hh:mm and does not enclose in quotes if value is a DateTime' do
+          value = DateTime.now
+          active_query.where('Field__c < :field', field: value)
+          expect(active_query.to_s).to end_with("(Field__c < #{value.iso8601})")
+        end
+
+        it 'formats as YYYY-MM-DDThh:mm:ss-hh:mm and does not enclose in quotes if value is a Time' do
+          value = Time.now
+          active_query.where('Field__c < :field', field: value)
+          expect(active_query.to_s).to end_with("(Field__c < #{value.iso8601})")
+        end
+
+        it 'formats as YYYY-MM-DD and does not enclose in quotes if value is a Date' do
+          value = Date.today
+          active_query.where('Field__c < :field', field: value)
+          expect(active_query.to_s).to end_with("(Field__c < #{value.iso8601})")
         end
 
         it 'accepts multiple bind parameters' do